fixing compiler warnings for 64-bit version
git-svn-id: svn://orca.physics.unc.edu/Orca64@11892 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/Ncd/Detector/NcdDetector.h
@@ -60,7 +60,7 @@
 - (NSMutableArray*) tubes;
 - (void) setTubes:(NSMutableArray*)newTubes;
 - (NcdTube*) tube:(int)index;
-- (int) numberOfTubes;
+- (NSUInteger) numberOfTubes;
 - (ORTimeRate*) shaperTotalRate;
 - (void) setShaperTotalRate:(ORTimeRate*)newShaperTotalRate;
 - (ORTimeRate*) muxTotalRate;
@@ -81,8 +81,8 @@
 - (id)loadWithCoder:(NSCoder*)decoder;
 - (void)saveWithCoder:(NSCoder*)encoder;
 
-- (void) removeTubeAtIndex:(int)index;
-- (void) addTube:(NcdTube*)aTube atIndex:(int)index;
+- (void) removeTubeAtIndex:(NSUInteger)index;
+- (void) addTube:(NcdTube*)aTube atIndex:(NSUInteger)index;
 - (float) shaperRate;
 - (float) muxRate;
