+++ Source/Objects/Data Processing/Histogram Objects/BaseClasses/ORDataController.m
@@ -31,7 +31,7 @@
 @end
 #endif
 
-int windowSort(id w1, id w2, void *context) { return [[w2 title] compare:[w1 title]]; }
+NSInteger windowSort(id w1, id w2, void *context) { return [[w2 title] compare:[w1 title]]; }
 
 
 @implementation ORDataController
